requirements.txt
anyio==3.4.0
asgiref==3.4.1
astroid==2.9.3
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
coverage==6.3.1
cryptography==36.0.1
ecdsa==0.17.0
ecs-logging==1.1.0
elastic-apm==6.7.2
fastapi==0.71.0
greenlet==1.1.2
h11==0.12.0
httptools==0.3.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
metrics==0.3.3
packaging==21.3
passlib==1.7.4
pathlib2==2.3.6
pathspec==0.5.5
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.12.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
Pygments==2.2.0
pylint==2.12.2
pylint-pydantic==0.1.4
pyparsing==3.0.6
pytest==6.2.5
pytest-cov==3.0.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
starlette-prometheus==0.9.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wrapt==1.13.3
